+++ Example/Sources/AppDelegate.swift
@@ -19,6 +19,7 @@ final class AppDelegate: UIResponder, UIApplicationDelegate {
     
     func application(_ application: UIApplication,
                      didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        ClientLogger.iconEnabled = true
         return true
     }