+++ Sources/Client/Client/Client.swift
@@ -163,12 +163,12 @@ public final class Client: Uploader {
         self.defaultURLSessionConfiguration = defaultURLSessionConfiguration
 
         if !apiKey.isEmpty, logOptions.isEnabled {
-            ClientLogger.logger("💬", "", .info, "Stream Chat v.\(Environment.version)")
-            ClientLogger.logger("🔑", "", .info, apiKey)
-            ClientLogger.logger("🔗", "", .info, baseURL.description)
+            ClientLogger.log("💬", "", .info, "Stream Chat v.\(Environment.version)")
+            ClientLogger.log("🔑", "", .info, apiKey)
+            ClientLogger.log("🔗", "", .info, baseURL.description)
             
             if let database = database {
-                ClientLogger.logger("💽", "", .info, "\(database.self)")
+                ClientLogger.log("💽", "", .info, "\(database.self)")
             }
         }